+++ lib/bit_array.dart
@@ -7,3 +7,5 @@ import 'dart:typed_data';
 part 'src/bit_array.dart';
 part 'src/bit_counter.dart';
 part 'src/bit_set.dart';
+part 'src/composite_counter.dart';
+part 'src/composite_set.dart';