+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - `asUint64Iterable()` added to provide interface-level compatibility for future compressed bit arrays.
 - `BitSet` to have a non-bitarray backed interface (eg. compressed structures).
+- `CompositeSet` and `CompositeCounter` for arbitrary-large compressed bitmap operations.
 
 ## 1.0.0